refactor: move AssetUploadButton to separate component and reset file input by rerendering
