+++ src/script/page/message-list/InputBarControls/ControlButtons.tsx
@@ -23,6 +23,7 @@ import MessageTimerButton from '../MessageTimerButton';
 import {t} from 'Util/LocalizerUtil';
 import {Conversation} from 'src/script/entity/Conversation';
 import {Config} from '../../../Config';
+import {AssetUploadButton} from '../AssetUploadButton';
 
 export type ControlButtonsProps = {
   input: string;
@@ -54,21 +55,10 @@ const ControlButtons: React.FC<ControlButtonsProps> = ({
   onClickGif,
 }) => {
   const acceptedImageTypes = Config.getConfig().ALLOWED_IMAGE_TYPES.join(',');
-  const acceptedFileTypes = Config.getConfig().FEATURE.ALLOWED_FILE_UPLOAD_EXTENSIONS.join(',');
 
   const imageRef = useRef<HTMLInputElement>(null!);
-  const fileRef = useRef<HTMLInputElement>(null!);
 
   const pingTooltip = t('tooltipConversationPing');
-
-  const handleFileChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const {files} = event.target;
-    if (!files) {
-      return;
-    }
-    onSelectFiles(Array.from(files));
-    event.target.value = '';
-  };
 
   if (isEditing) {
     return (
@@ -128,25 +118,7 @@ const ControlButtons: React.FC<ControlButtonsProps> = ({
             </li>
 
             <li>
-              <button
-                type="button"
-                aria-label={t('tooltipConversationFile')}
-                title={t('tooltipConversationFile')}
-                className="conversation-button controls-right-button no-radius file-button"
-                onClick={() => fileRef.current?.click()}
-                data-uie-name="do-share-file"
-              >
-                <Icon.Attachment />
-
-                <input
-                  ref={fileRef}
-                  accept={acceptedFileTypes ?? null}
-                  id="conversation-input-bar-files"
-                  tabIndex={-1}
-                  onChange={handleFileChange}
-                  type="file"
-                />
-              </button>
+              <AssetUploadButton onSelectFiles={onSelectFiles} />
             </li>
           </>
         )}
